feat(ui): use pie-catalog-client ui library
This pr uses pie-catalog-client for `pie info` and `pie serve`

Note that there is still some outstanding work on tidying up the app logic but this will be the focus of #97.
The new apps (info and item) no longer use `BaseApp` and it seems like we'll probably want to do away with it.

Closes #98