feat(install): Change underlying install model.
The underlying installation model has changed significantly. This is to add support for scoped packages, but also to detach the element key in the `config.json` from the package name that is installed.

This gives greater flexibility in how the packages are pulled in.

The new flow is for the `elements` map in the `config.json` to be incrementally procesed where we analyse if the element needs to be installed, whether it's local/remote and whether it's a pie.
Actions are triggered from here to do the installation and the information is returned as an array of `PieBuildInfo` types that contain all the relevant information.